[Chore] move `ITraitChangeEvent` to `listWidgetTrait.ts`
